requirements.txt
asgiref==3.7.2
asttokens==2.4.0
autocommand==2.2.2
backcall==0.2.0
billiard==4.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
cryptography==41.0.4
Cython==3.0.2
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-guardian==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-guardian==0.3.0
drf-extensions==0.7.1
drf-writable-nested==0.7.0
evdev==1.6.1
executing==1.2.0
fastjsonschema==2.18.0
hid-parser==0.0.3
idna==3.4
inflect==7.0.0
ipython==8.15.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
jedi==0.19.0
kombu==5.3.2
matplotlib-inline==0.1.6
model-bakery==1.15.0
more-itertools==10.1.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.16.1
PyGObject==3.44.1
PyJWT==2.8.0
python3-openid==3.2.0
python-dateutil==2.8.2
python-xlib==0.33
pytz==2023.3.post1
pyudev==0.24.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
solaar==1.1.9
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.10.0
trove-classifiers==2023.8.8
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
vine==5.0.0
wcwidth==0.2.6
yt-dlp==2023.7.6
